Refactor view names
