+++ letter_templates/urls.py
@@ -1,6 +1,8 @@
 from django.urls import path
 
-from letter_templates.views import manage, generator
+import letter_templates.views.edit
+import letter_templates.views.letter_paragraphs
+from letter_templates.views import view, create
 
 app_name = 'letter_templates'
 
@@ -8,22 +10,22 @@ urlpatterns = [
     # Manage letter templates
 
     # ex: /<uuid:pk>/letter-templates/
-    path('', manage.LetterTemplatesList.as_view(), name='letter_templates'),
+    path('', view.LetterTemplatesList.as_view(), name='letter_templates'),
     # ex: /<uuid:pk>/letter-templates/
-    path('<uuid:pk>/', manage.LetterTemplateDetail.as_view(), name='letter_template'),
+    path('<uuid:pk>/', view.LetterTemplateDetail.as_view(), name='letter_template'),
     # ex: /<uuid:pk>/letter-templates/edit/
-    path('<uuid:pk>/edit/', manage.LetterTemplateEdit.as_view(), name='edit'),
+    path('<uuid:pk>/edit/', letter_templates.views.edit.LetterTemplateEdit.as_view(), name='edit'),
     # ex: /<uuid:pk>/letter-templates/edit/
-    path('<uuid:pk>/edit-letter-paragraphs/', manage.LetterTemplateEditLetterParagraphs.as_view(), name='edit_letter_paragraphs'),
+    path('<uuid:pk>/edit-letter-paragraphs/', letter_templates.views.edit.LetterTemplateEditLetterParagraphs.as_view(), name='edit_letter_paragraphs'),
 
     # Create letter templates
 
     # ex: /letter-templates/add/
-    path('add/', generator.Add.as_view(), name='add'),
+    path('add/', create.Add.as_view(), name='add'),
     # ex: /letter-templates/add/letter-paragraphs/
-    path('add/letter-paragraphs/', generator.LetterParagraphs.as_view(), name='letter_paragraphs'),
+    path('add/letter-paragraphs/', letter_templates.views.letter_paragraphs.LetterParagraphs.as_view(), name='letter_paragraphs'),
     # ex: /letter-templates/preview/
-    path('preview/', generator.Preview.as_view(), name='preview'),
+    path('preview/', create.Preview.as_view(), name='preview'),
     # ex: /letter-templates/create/
-    path('create/', generator.Create.as_view(), name='create'),
+    path('create/', create.Create.as_view(), name='create'),
 ]
