+++ letter_templates/views/edit.py
@@ -1,35 +1,12 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import redirect, render
 from django.urls import reverse
 from django.views.generic import TemplateView
 from lite_forms.generators import form_page
 from lite_forms.submitters import submit_single_form
 
-from letter_templates import helpers
 from letter_templates.forms import edit_letter_template
-from letter_templates.services import get_letter_paragraphs, get_letter_templates, get_letter_template, \
-    put_letter_template
+from letter_templates.services import get_letter_template, put_letter_template, get_letter_paragraphs
 from picklists.services import get_picklists
-
-
-class LetterTemplatesList(TemplateView):
-    def get(self, request, **kwargs):
-        context = {
-            'letter_templates': get_letter_templates(request)
-        }
-        return render(request, 'letter_templates/letter_templates.html', context)
-
-
-class LetterTemplateDetail(TemplateView):
-    def get(self, request, **kwargs):
-        letter_template_id = str(kwargs['pk'])
-        letter_template = get_letter_template(request, letter_template_id)
-        letter_paragraphs = get_letter_paragraphs(request, letter_template['letter_paragraphs'])
-
-        context = {
-            'letter_template': letter_template,
-            'preview': helpers.generate_preview(letter_template['layout']['id'], letter_paragraphs)
-        }
-        return render(request, 'letter_templates/letter_template.html', context)
 
 
 class LetterTemplateEdit(TemplateView):
